HD21823 / HD22069 browse entry cleanup speed up

+++ KCLS/sql/upgrade/deploy/0010-metabib-browse-cache-fix.sql
@@ -37,14 +37,26 @@ BEGIN
     -- At this point, browse entries mapped to the deleted bibs/auths
     -- have already been removed.
     IF TG_OP = 'DELETE' THEN
-        SELECT INTO keep_entry EXISTS (
-            SELECT TRUE FROM metabib.browse_entry_def_map WHERE entry = entry_id
-            UNION
-            SELECT TRUE FROM metabib.browse_entry_simple_heading_map WHERE entry = entry_id
-        );
-        IF NOT keep_entry THEN
-            DELETE FROM metabib.browse_entry WHERE id = entry_id;
+
+        -- start with a faster query to see if any fields are known
+        -- reference the browse entry we're potentially removing.
+        PERFORM TRUE FROM metabib.browse_entry
+        WHERE id = entry_id AND ARRAY_LENGTH(metabib_fields_cache, 1) > 0;
+        
+        IF NOT FOUND THEN
+            -- If no obvious links are found, check for specific map
+            -- entries. These can exist e.g. if an authority field does 
+            -- link to a bib field and thus config.metabib_field.
+            SELECT INTO keep_entry EXISTS (
+                SELECT TRUE FROM metabib.browse_entry_def_map WHERE entry = entry_id
+                UNION
+                SELECT TRUE FROM metabib.browse_entry_simple_heading_map WHERE entry = entry_id
+            );
+            IF NOT keep_entry THEN
+                DELETE FROM metabib.browse_entry WHERE id = entry_id;
+            END IF;
         END IF;
+
     END IF;
 
     RETURN NEW;
